build(webpack): remove redundant environment check
Webpack v4 introduced a new `mode` setting, which - amongst other
things - automates the definition of the NODE_ENV environment variable.
We already had the `mode` settings set in our webpack configs so this
change is just to clean up some of the redundant code that is handled
automatically by this setting.

See:
 - https://webpack.js.org/concepts/mode/