refactor: rework containers and routes
Refactor Containers and Routes in order to make the code a little more
readable, reduce duplication, and prepare the codebase for integration
of the new Home screen.

- Break out the Root container into several smaller containers that are
responsible for connecting themselves to the Redux store.

- Use routing for navigation between primary containers

- Unbundle the `Syncing` Component from the Onboarding section and hook
up to router so that it can be access by navigating to `/syncing' (we
want to be able to load existing wallets and initiate syncing outside of
the context of the onboarding process)

- Reduce code duplication
  - Create `TitleBar` component and implement once at the top level
  - Create `withLoading` HOC and use to display the loading screen when
  needed.